Refactor sorting algorithm classes to improve readability and add missing Javadoc
