+++ src/main/java/com/github/netmastermichael/classic_algorithms_suite/SortingAlgorithms/SortingAlgorithmOperation.java
@@ -1,5 +1,10 @@
 package com.github.netmastermichael.classic_algorithms_suite.SortingAlgorithms;
 
+/**
+ * SortingAlgorithmOperation is an enum that represents the possible types of
+ * operations that a sorting algorithm may perform. Intended for use in
+ * ManualSorter.
+ */
 public enum SortingAlgorithmOperation {
 	SWAP, COMPARE;
 }
